Proper error handling, Fake City, Bootstrap 4 fix
Error handling wasn't working properly, it was intended to display an error message from the API.
But since the API returns a 404 on a bad request, the error function was being called, not the success.
Now everything works as intended.
Added comments to explain.
Also fixed table not being hidden because was using Bootstrap 3 class.